requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
asttokens==2.2.1
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
exceptiongroup==1.0.4
execnet==1.9.0
executing==1.2.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.0.0
iniconfig==1.1.1
ipykernel==6.21.2
ipython==8.11.0
jedi==0.18.2
jsonpath-ng==1.6.0
jupyter_client==8.0.3
jupyter_core==5.2.0
lxml==5.1.0
matplotlib-inline==0.1.6
MouseInfo==0.1.3
nest-asyncio==1.5.6
oauthlib==3.2.2
outcome==1.2.0
packaging==21.3
parameterized==0.8.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.0.0
pluggy==1.0.0
ply==3.11
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.4
pure-eval==0.2.2
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.14.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pytest==7.2.0
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-xdist==3.3.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
pytweening==1.0.4
pyzmq==25.1.2
requests==2.28.1
requests-oauthlib==1.3.1
retry==0.9.2
rsa==4.9
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
stack-data==0.6.2
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.6
webdriver-manager==4.0.0
wsproto==1.2.0
zipp==3.15.0
